acquire_.c: changed int to size_t. gssapip_.h: added prototype for kg_seal_size. k5seal.c: 16/32 bit mismatch and removed unused variables. seal.c: 16/32 bit mismatch. sign.c: 16/32 bit mismatch. ser_sctx.c: added prototypes for all functions since they get 	assigned into a structure that has been prototyped.
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@6833 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/gssapi/krb5/seal.c
@@ -55,7 +55,7 @@ krb5_gss_wrap(context, minor_status, context_handle, conf_req_flag,
     gss_buffer_t	output_message_buffer;
 {
     return(kg_seal(context, minor_status, context_handle, conf_req_flag,
-		   qop_req, input_message_buffer, conf_state,
+		   (int) qop_req, input_message_buffer, conf_state,
 		   output_message_buffer, KG_TOK_WRAP_MSG));
 }
